New AddRequirementCard design and UX
